+++ src/mlx_window.c
@@ -6,7 +6,7 @@
 /*   By: W2wizard <w2wizzard@gmail.com>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/02/08 01:14:59 by W2wizard      #+#    #+#                 */
-/*   Updated: 2022/07/05 16:03:41 by jobvan-d      ########   odam.nl         */
+/*   Updated: 2022/06/29 15:34:25 by lde-la-h      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
  * Recalculate the view projection matrix, used by images for screen pos
  * Reference: https://bit.ly/3KuHOu1 (Matrix View Projection)
  */
-void mlx_update_matrix(const mlx_t* mlx, uint32_t width, uint32_t height)
+void mlx_update_matrix(const mlx_t* mlx, int32_t width, int32_t height)
 {
 	const mlx_ctx_t* mlxctx = mlx->context;
 	const float depth = mlxctx->zdepth;
@@ -47,7 +47,7 @@ static void mlx_resize_callback(GLFWwindow* window, int32_t width, int32_t heigh
 	const mlx_ctx_t* mlxctx = mlx->context;
 
 	if (mlxctx->resize_hook.func)
-		mlxctx->resize_hook.func((uint32_t)width, (uint32_t)height, mlxctx->resize_hook.param);
+		mlxctx->resize_hook.func(width, height, mlxctx->resize_hook.param);
 }
 
 static void mlx_close_callback(GLFWwindow* window)
@@ -106,13 +106,13 @@ void mlx_get_window_pos(mlx_t* mlx, int32_t* xpos, int32_t* ypos)
 	glfwGetWindowPos(mlx->window, xpos, ypos);
 }
 
-void mlx_set_window_size(mlx_t* mlx, uint32_t new_width, uint32_t new_height)
+void mlx_set_window_size(mlx_t* mlx, int32_t new_width, int32_t new_height)
 {
 	MLX_NONNULL(mlx);
 
 	mlx->width = new_width;
 	mlx->height = new_height;
-	glfwSetWindowSize(mlx->window, (int)new_width, (int)new_height);
+	glfwSetWindowSize(mlx->window, new_width, new_height);
 }
 
 void mlx_set_window_limit(mlx_t* mlx, int32_t min_w, int32_t min_h, int32_t max_w, int32_t max_h)
